@@ -36,19 +36,23 @@ dependencies:
   image_picker: ^0.6.1+8
   dio: ^3.0.9
   shared_preferences: ^2.0.5
-  charts_flutter: ^0.9.0
+#  charts_flutter: ^0.11.0
+  charts_flutter:
+    git:
+      url: git://github.com/google/charts.git
+      path: charts_flutter
   flutter_slidable: ^0.5.4
   flutter_html: ^0.11.0
-  barcode_scan: ^1.0.0
+  barcode_scan: ^3.0.1
   share: ^0.6.2+4
   path_provider: ^2.0.1
-  open_file: ^2.0.3
-  flutter_document_picker: ^3.0.0
+  open_file: ^3.2.1
+  flutter_document_picker: ^4.0.1
   android_intent: ^2.0.0
   chewie: ^0.9.8+1
   event_bus: ^1.1.0
-  fluttertoast: ^3.1.3
-  flutter_webview_plugin: ^0.3.8
+  fluttertoast: ^8.0.8
+  webview_flutter: ^2.0.10
   web_socket_channel: ^1.1.0
   fcharts: ^0.0.10
   sqflite: ^1.1.7+1
