@@ -89,7 +89,7 @@ class _IndexSlidersPageState extends State<IndexSlidersPage> {
               valueIndicatorColor: Colors.deepPurpleAccent,
               thumbShape: _CustomThumbShape(),
               valueIndicatorShape: _CustomValueIndicatorShape(),
-              valueIndicatorTextStyle: theme.accentTextTheme.body2.copyWith(color: Colors.black87),
+              valueIndicatorTextStyle: theme.accentTextTheme.bodyText1.copyWith(color: Colors.black87),
             ),
             child: Slider(
               value: _value3,
